add rec_resnet45
